Refactor backend init, expose systemTheme property
- Refactor backend to use class-based structure for Tauri and Web
- Move all backend types into backend.ts
- Expose systemTheme as a readable store on IBackend
- Update Tauri & Web implementations, supply systemTheme
- Refactor theme utility to leverage backend systemTheme property
