+++ apps/desktop/src/lib/utils/theme.ts
@@ -1,25 +1,13 @@
-import { getCurrentWindow, Window, type Theme } from '@tauri-apps/api/window';
 import { type Writable } from 'svelte/store';
+import type { IBackend } from '$lib/backend';
 import type { Settings } from '$lib/settings/userSettings';
-
-let appWindow: Window | undefined;
-if (import.meta.env.VITE_BUILD_TARGET === 'web') {
-	// TODO: Implement electron alternative
-} else {
-	appWindow = getCurrentWindow();
-}
 
 let systemTheme: string | null;
 let selectedTheme: string | undefined;
 
-export function initTheme(userSettings: Writable<Settings>) {
-	if (!appWindow) return;
-	appWindow.theme().then((value: Theme | null) => {
-		systemTheme = value;
-		updateDom();
-	});
-	appWindow.onThemeChanged((e) => {
-		systemTheme = e.payload;
+export function initTheme(userSettings: Writable<Settings>, backend: IBackend) {
+	backend.systemTheme.subscribe((theme) => {
+		systemTheme = theme;
 		updateDom();
 	});
 	userSettings.subscribe((s) => {
